BigCommerce Webhook app using Express.js
This app is designed to watch for BigCommerce store changes using the BigCommerce API in Express.js and produce a resulting HTML page showing information about the operation.

This app includes only the app files and ignores the node_modules directory since these files come with Express and have not been altered.